+++ src/main/java/com/bokmcdok/butterflies/client/gui/screens/ButterflyBookScreen.java
@@ -1,7 +1,8 @@
 package com.bokmcdok.butterflies.client.gui.screens;
 
+import com.bokmcdok.butterflies.butterfly_data.ButterflyRegistry;
 import com.bokmcdok.butterflies.client.texture.ButterflyTextures;
-import com.bokmcdok.butterflies.world.ButterflyData;
+import com.bokmcdok.butterflies.butterfly_data.ButterflyData;
 import com.bokmcdok.butterflies.world.CompoundTagId;
 import net.minecraft.client.GameNarrator;
 import net.minecraft.client.gui.GuiGraphics;
@@ -198,7 +199,7 @@ public class ButterflyBookScreen extends Screen {
         this.cachedPage = this.currentPage;
         if (this.cachedPage % 2 == 0) {
             int butterflyIndex = bookAccess.getButterflyIndex(cachedPage);
-            ButterflyData data = ButterflyData.getEntry(butterflyIndex);
+            ButterflyData data = ButterflyRegistry.getEntry(butterflyIndex);
             if (data != null) {
                 guiGraphics.blit(data.getScrollTexture(), i, 2, 0, 0, 192, 192);
             }
@@ -398,7 +399,7 @@ public class ButterflyBookScreen extends Screen {
         public FormattedText getPageRaw(int page) {
             if (pages != null) {
                 int butterflyIndex = pages.getInt((page - 1) / 2);
-                return ButterflyData.getFormattedButterflyData(butterflyIndex);
+                return ButterflyTextFormatter.getFormattedButterflyData(butterflyIndex);
             }
 
             return null;